+++ smartcontract/sdk/rs/src/commands/multicastgroup/allowlist/publisher/add.rs
@@ -27,7 +27,7 @@ impl AddMulticastGroupPubAllowlistCommand {
 
         let (globalstate_pubkey, _) = get_globalstate_pda(&client.get_program_id());
 
-        client.execute_transaction(
+        client.execute_authorized_transaction(
             DoubleZeroInstruction::AddMulticastGroupPubAllowlist(
                 AddMulticastGroupPubAllowlistArgs {
                     client_ip: self.client_ip,
@@ -100,7 +100,7 @@ mod tests {
                 Ok(map)
             });
         client
-            .expect_execute_transaction()
+            .expect_execute_authorized_transaction()
             .with(
                 predicate::eq(DoubleZeroInstruction::AddMulticastGroupPubAllowlist(
                     AddMulticastGroupPubAllowlistArgs {