+++ smartcontract/sdk/rs/src/commands/multicastgroup/allowlist/subscriber/add.rs
@@ -29,7 +29,7 @@ impl AddMulticastGroupSubAllowlistCommand {
 
         let (globalstate_pubkey, _) = get_globalstate_pda(&client.get_program_id());
 
-        client.execute_transaction(
+        client.execute_authorized_transaction(
             DoubleZeroInstruction::AddMulticastGroupSubAllowlist(
                 AddMulticastGroupSubAllowlistArgs {
                     client_ip: self.client_ip,
@@ -102,7 +102,7 @@ mod tests {
                 Ok(map)
             });
         client
-            .expect_execute_transaction()
+            .expect_execute_authorized_transaction()
             .with(
                 predicate::eq(DoubleZeroInstruction::AddMulticastGroupSubAllowlist(
                     AddMulticastGroupSubAllowlistArgs {