+++ smartcontract/sdk/rs/src/commands/multicastgroup/delete.rs
@@ -36,7 +36,7 @@ impl DeleteMulticastGroupCommand {
             AccountMeta::new(mgroup.owner, false),
         ];
 
-        client.execute_transaction(
+        client.execute_authorized_transaction(
             DoubleZeroInstruction::DeleteMulticastGroup(MulticastGroupDeleteArgs {
                 use_onchain_deallocation: true,
             }),
@@ -101,7 +101,7 @@ mod tests {
             .returning(move |_| Ok(AccountData::MulticastGroup(mgroup_cloned.clone())));
 
         client
-            .expect_execute_transaction()
+            .expect_execute_authorized_transaction()
             .with(
                 predicate::eq(DoubleZeroInstruction::DeleteMulticastGroup(
                     MulticastGroupDeleteArgs {